requirements.txt
astral
click
dateparser
loguru
Pillow
Requests
setuptools
